Refactored FilterDropdown into multiple components
