+++ front/src/modules/filters-and-sorts/states/isFilterDropdownOperandSelectUnfoldedScopedState.ts
@@ -1,0 +1,9 @@
+import { atomFamily } from 'recoil';
+
+export const isFilterDropdownOperandSelectUnfoldedScopedState = atomFamily<
+  boolean,
+  string
+>({
+  key: 'isFilterDropdownOperandSelectUnfoldedScopedState',
+  default: false,
+});
