+++ front/src/modules/ui/components/table/table-header/FilterDropdownButton.tsx
@@ -1,42 +1,36 @@
-import { ChangeEvent, useCallback, useState } from 'react';
-import styled from '@emotion/styled';
+import { useCallback, useState } from 'react';
 import { useRecoilState } from 'recoil';
 
-import { useRemoveSelectedFilter } from '@/filters-and-sorts/hooks/useRemoveSelectedFilter';
-import { useUpsertSelectedFilter } from '@/filters-and-sorts/hooks/useUpsertSelectedFilter';
-import { availableFiltersScopedState } from '@/filters-and-sorts/states/availableFiltersScopedState';
 import { filterSearchInputScopedState } from '@/filters-and-sorts/states/filterSearchInputScopedState';
+import { isFilterDropdownOperandSelectUnfoldedScopedState } from '@/filters-and-sorts/states/isFilterDropdownOperandSelectUnfoldedScopedState';
 import { selectedFilterInDropdownScopedState } from '@/filters-and-sorts/states/selectedFilterInDropdownScopedState';
 import { selectedFiltersScopedState } from '@/filters-and-sorts/states/selectedFiltersScopedState';
 import { selectedOperandInDropdownScopedState } from '@/filters-and-sorts/states/selectedOperandInDropdownScopedState';
-import { getOperandLabel } from '@/filters-and-sorts/utils/getOperandLabel';
-import { getOperandsForFilterType } from '@/filters-and-sorts/utils/getOperandsForFilterType';
 import { captureHotkeyTypeInFocusState } from '@/hotkeys/states/captureHotkeyTypeInFocusState';
-import { RecoilScope } from '@/recoil-scope/components/RecoilScope';
 import { useRecoilScopedState } from '@/recoil-scope/hooks/useRecoilScopedState';
-import { useRecoilScopedValue } from '@/recoil-scope/hooks/useRecoilScopedValue';
 import { TableContext } from '@/ui/tables/states/TableContext';
 
-import DatePicker from '../../form/DatePicker';
-import { DropdownMenuItemContainer } from '../../menu/DropdownMenuItemContainer';
-import { DropdownMenuSelectableItem } from '../../menu/DropdownMenuSelectableItem';
-import { DropdownMenuSeparator } from '../../menu/DropdownMenuSeparator';
-
 import DropdownButton from './DropdownButton';
+import { FilterDropdownDateSearchInput } from './FilterDropdownDateSearchInput';
+import { FilterDropdownEntitySearchInput } from './FilterDropdownEntitySearchInput';
+import { FilterDropdownEntitySelect } from './FilterDropdownEntitySelect';
+import { FilterDropdownFilterSelect } from './FilterDropdownFilterSelect';
+import { FilterDropdownNumberSearchInput } from './FilterDropdownNumberSearchInput';
+import { FilterDropdownOperandButton } from './FilterDropdownOperandButton';
+import { FilterDropdownOperandSelect } from './FilterDropdownOperandSelect';
+import { FilterDropdownTextSearchInput } from './FilterDropdownTextSearchInput';
 
 export function FilterDropdownButton() {
-  const availableFilters = useRecoilScopedValue(
-    availableFiltersScopedState,
-    TableContext,
-  );
-
   const [isUnfolded, setIsUnfolded] = useState(false);
   const [, setCaptureHotkeyTypeInFocus] = useRecoilState(
     captureHotkeyTypeInFocusState,
   );
 
   const [isOperandSelectionUnfolded, setIsOperandSelectionUnfolded] =
-    useState(false);
+    useRecoilScopedState(
+      isFilterDropdownOperandSelectUnfoldedScopedState,
+      TableContext,
+    );
 
   const [selectedFilterInDropdown, setSelectedFilterInDropdown] =
     useRecoilScopedState(selectedFilterInDropdownScopedState, TableContext);
@@ -54,10 +48,6 @@ export function FilterDropdownButton() {
   const [selectedOperandInDropdown, setSelectedOperandInDropdown] =
     useRecoilScopedState(selectedOperandInDropdownScopedState, TableContext);
 
-  const operandsForFilterType = getOperandsForFilterType(
-    selectedFilterInDropdown?.type,
-  );
-
   const resetState = useCallback(() => {
     setIsOperandSelectionUnfolded(false);
     setSelectedFilterInDropdown(null);
@@ -67,10 +57,8 @@ export function FilterDropdownButton() {
     setSelectedFilterInDropdown,
     setSelectedOperandInDropdown,
     setFilterSearchInput,
+    setIsOperandSelectionUnfolded,
   ]);
-
-  const upsertSelectedFilter = useUpsertSelectedFilter();
-  const removeSelectedFilter = useRemoveSelectedFilter();
 
   const isFilterSelected = (selectedFilters?.length ?? 0) > 0;
 
@@ -82,103 +70,33 @@ export function FilterDropdownButton() {
       setIsUnfolded={setIsUnfolded}
       resetState={resetState}
     >
-      {selectedFilterInDropdown && selectedOperandInDropdown ? (
-        isOperandSelectionUnfolded ? (
-          operandsForFilterType.map((filterOperand, index) => (
-            <DropdownButton.StyledDropdownItem
-              key={`select-filter-operand-${index}`}
-              onClick={() => {
-                setSelectedOperandInDropdown(filterOperand);
-                setIsOperandSelectionUnfolded(false);
-              }}
-            >
-              {getOperandLabel(filterOperand)}
-            </DropdownButton.StyledDropdownItem>
-          ))
-        ) : (
+      {!selectedFilterInDropdown ? (
+        <FilterDropdownFilterSelect />
+      ) : isOperandSelectionUnfolded ? (
+        <FilterDropdownOperandSelect />
+      ) : (
+        selectedOperandInDropdown && (
           <>
-            <DropdownButton.StyledDropdownTopOption
-              key={'selected-filter-operand'}
-              onClick={() => setIsOperandSelectionUnfolded(true)}
-            >
-              {getOperandLabel(selectedOperandInDropdown)}
-              <DropdownButton.StyledDropdownTopOptionAngleDown />
-            </DropdownButton.StyledDropdownTopOption>
+            <FilterDropdownOperandButton />
             <DropdownButton.StyledSearchField autoFocus key={'search-filter'}>
-              {(selectedFilterInDropdown.type === 'text' ||
-                selectedFilterInDropdown.type === 'entity') && (
-                <input
-                  type="text"
-                  placeholder={selectedFilterInDropdown.label}
-                  onChange={(event: ChangeEvent<HTMLInputElement>) => {
-                    if (selectedFilterInDropdown.type === 'entity') {
-                      setFilterSearchInput(event.target.value);
-                    }
-
-                    if (selectedFilterInDropdown.type === 'text') {
-                      if (event.target.value === '') {
-                        removeSelectedFilter(selectedFilterInDropdown.field);
-                      } else {
-                        upsertSelectedFilter({
-                          field: selectedFilterInDropdown.field,
-                          type: selectedFilterInDropdown.type,
-                          value: event.target.value,
-                          operand: selectedOperandInDropdown,
-                          displayValue: event.target.value,
-                        });
-                      }
-                    }
-                  }}
-                />
+              {selectedFilterInDropdown.type === 'text' && (
+                <FilterDropdownTextSearchInput />
+              )}
+              {selectedFilterInDropdown.type === 'number' && (
+                <FilterDropdownNumberSearchInput />
               )}
               {selectedFilterInDropdown.type === 'date' && (
-                <DatePicker
-                  date={new Date()}
-                  onChangeHandler={(date) => {
-                    upsertSelectedFilter({
-                      field: selectedFilterInDropdown.field,
-                      type: selectedFilterInDropdown.type,
-                      value: date.toISOString(),
-                      operand: selectedOperandInDropdown,
-                      displayValue: date.toLocaleDateString(),
-                    });
-                  }}
-                  customInput={<></>}
-                  customCalendarContainer={styled.div`
-                    top: -10px;
-                  `}
-                />
+                <FilterDropdownDateSearchInput />
+              )}
+              {selectedFilterInDropdown.type === 'entity' && (
+                <FilterDropdownEntitySearchInput />
               )}
             </DropdownButton.StyledSearchField>
-            <DropdownMenuSeparator />
             {selectedFilterInDropdown.type === 'entity' && (
-              <RecoilScope>
-                {selectedFilterInDropdown.searchSelectComponent}
-              </RecoilScope>
+              <FilterDropdownEntitySelect />
             )}
           </>
         )
-      ) : (
-        <DropdownMenuItemContainer>
-          {availableFilters.map((filter, index) => (
-            <DropdownMenuSelectableItem
-              key={`select-filter-${index}`}
-              onClick={() => {
-                setSelectedFilterInDropdown(filter);
-                setSelectedOperandInDropdown(
-                  getOperandsForFilterType(filter.type)?.[0],
-                );
-
-                setFilterSearchInput('');
-              }}
-            >
-              <DropdownButton.StyledIcon>
-                {filter.icon}
-              </DropdownButton.StyledIcon>
-              {filter.label}
-            </DropdownMenuSelectableItem>
-          ))}
-        </DropdownMenuItemContainer>
       )}
     </DropdownButton>
   );
