+++ front/src/modules/ui/components/table/table-header/FilterDropdownOperandSelect.tsx
@@ -1,0 +1,53 @@
+import { isFilterDropdownOperandSelectUnfoldedScopedState } from '@/filters-and-sorts/states/isFilterDropdownOperandSelectUnfoldedScopedState';
+import { selectedFilterInDropdownScopedState } from '@/filters-and-sorts/states/selectedFilterInDropdownScopedState';
+import { selectedOperandInDropdownScopedState } from '@/filters-and-sorts/states/selectedOperandInDropdownScopedState';
+import { getOperandLabel } from '@/filters-and-sorts/utils/getOperandLabel';
+import { getOperandsForFilterType } from '@/filters-and-sorts/utils/getOperandsForFilterType';
+import { useRecoilScopedState } from '@/recoil-scope/hooks/useRecoilScopedState';
+import { TableContext } from '@/ui/tables/states/TableContext';
+
+import { DropdownMenuItemContainer } from '../../menu/DropdownMenuItemContainer';
+
+import DropdownButton from './DropdownButton';
+
+export function FilterDropdownOperandSelect() {
+  const [selectedFilterInDropdown] = useRecoilScopedState(
+    selectedFilterInDropdownScopedState,
+    TableContext,
+  );
+
+  const [, setSelectedOperandInDropdown] = useRecoilScopedState(
+    selectedOperandInDropdownScopedState,
+    TableContext,
+  );
+
+  const operandsForFilterType = getOperandsForFilterType(
+    selectedFilterInDropdown?.type,
+  );
+
+  const [isOperandSelectionUnfolded, setIsOperandSelectionUnfolded] =
+    useRecoilScopedState(
+      isFilterDropdownOperandSelectUnfoldedScopedState,
+      TableContext,
+    );
+
+  if (!isOperandSelectionUnfolded) {
+    return <></>;
+  }
+
+  return (
+    <DropdownMenuItemContainer>
+      {operandsForFilterType.map((filterOperand, index) => (
+        <DropdownButton.StyledDropdownItem
+          key={`select-filter-operand-${index}`}
+          onClick={() => {
+            setSelectedOperandInDropdown(filterOperand);
+            setIsOperandSelectionUnfolded(false);
+          }}
+        >
+          {getOperandLabel(filterOperand)}
+        </DropdownButton.StyledDropdownItem>
+      ))}
+    </DropdownMenuItemContainer>
+  );
+}
